requirements.txt
python-dotenv>=0.19.0
openai==1.57.0
pydantic>=2.0.0 